added signup and login code in the same file
