--- backend/src/controllers/loginController.js
@@ -1,34 +1,0 @@
-import prisma from '../prisma/client.js';
-import bcrypt from 'bcrypt';
-import jwt from 'jsonwebtoken';
-
-export const login = async (req, res) => {
-  const { email, password } = req.body;
-
-  try {
-    if (!email || !password) {
-      return res
-        .status(400)
-        .json({ message: 'Email and Password is required' });
-    }
-
-    const user = await prisma.user.findUnique({ where: { email } });
-    if (!user) {
-      return res.status(400).json({ message: 'Invalid credentials' });
-    }
-
-    const isPasswordCorrect = await bcrypt.compare(password, user.password);
-    if (!isPasswordCorrect) {
-      return res.status(400).json({ message: 'Invalid credentials' });
-    }
-
-    const token = jwt.sign({ userId: user.id }, process.env.JWT_SECRET, {
-      expiresIn: '7d',
-    });
-
-    res.status(200).json({message:"Login successfull!", token})
-  } catch (error) {
-    console.log('Error is login controller', error.message);
-    res.status(500).json({ message: 'Internal Server Error' });
-  }
-};
